signals-engine/requirements.txt
yfinance==0.2.43
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
pyyaml==6.0.2
scikit-learn==1.5.1
requests==2.32.3
vaderSentiment==3.3.2
